vms-f49: fix emutls control-object width (LLP64) — the veneer rung-4 crash root cause
The alpha emulated-TLS runtime typed the control-object fields as `unsigned long`
(emutls_word), which on the alpha-dec-vms LLP64 target is 32 BITS — but the cc1
emits the control with 64-bit .quad fields (.quad size; .quad align; .quad loc;
.quad templ). So struct __emutls_object packed size+align into the first 8 bytes
and put `loc` at offset 8 (the align field) instead of 16. __emutls_get_address
returned obj->loc = the align value (4), and the first __thread access on the
veneer's sys$create path (a vms_kif 'vms_bound_pid == getpid()' check) then
dereferenced 4 -> SIGSEGV at 0x4 — the rung-4 (vms-f49) blocker, pinned via a
local qemu-system-alpha boot + an IMGACT SIGSEGV-handler RA capture.

Fix: emutls_word -> unsigned long long (64-bit on every target), so loc lands at
offset 16 and templ at 24, matching the .quad emission. Same LLP64 bug class as
vms-1fc. Arch-scoped: the whole file is #if defined(__alpha__), so x86_64/aarch64
(musl TLS) and VAX (NetBSD) never compile it — the non-veneer + VAX/x86_64
controls stay byte-identical.

Verified locally: the v0=4 / gp-as-stack SIGSEGV is gone; the veneer image now
advances past the emutls point into the Files-11 path.

+++ src/libvmssys/arch/alpha/emutls_vms.c
@@ -46,12 +46,22 @@ extern void *memset(void *, int, unsigned long);
 /* The control object the compiler emits, in emission order. `word` is pointer-
  * width (8 bytes on this LP64-kernel/LLP64-compiler target: the .quad fields
  * above are 64-bit). */
-typedef unsigned long emutls_word;
+/* vms-f49 (rung 4): each control field the alpha-dec-vms cc1 emits is a `.quad`
+ * (64-bit) -- `__emutls_v.x: .quad size ; .quad align ; .quad loc ; .quad templ`
+ * (verified from generated assembly). emutls_word MUST therefore be 64-bit so
+ * `loc` lands at struct offset 16 and `templ` at 24, matching the emission. It
+ * was `unsigned long`, which on this LLP64 target is 32 BITS (same class as the
+ * vms-1fc width bug) -- that packed size+align into the first 8 bytes and put
+ * `loc` at offset 8 (the align field). __emutls_get_address then returned the
+ * align value (4) as the TLS pointer, and the first __thread access on the
+ * veneer's sys$create path dereferenced 4 -> SIGSEGV (the rung-4 blocker). Use
+ * `unsigned long long` (guaranteed 64-bit on every target) to match the .quad. */
+typedef unsigned long long emutls_word;
 struct __emutls_object {
-	emutls_word  size;   /* bytes of the __thread object            */
-	emutls_word  align;  /* required alignment                      */
-	void        *loc;    /* runtime-owned storage pointer (init 0)  */
-	void        *templ;  /* initializer image, or 0 => zero-init    */
+	emutls_word  size;   /* bytes of the __thread object  (.quad, offset 0)  */
+	emutls_word  align;  /* required alignment            (.quad, offset 8)  */
+	void        *loc;    /* runtime-owned storage pointer (.quad, offset 16) */
+	void        *templ;  /* initializer image, 0=>zero    (.quad, offset 24) */
 };
 
 void *__emutls_get_address(struct __emutls_object *obj)
